fix remaining cppcheck warnings in ecl package

+++ ecl/modules/eclPacker/src/eclPackerModule.cc
@@ -315,7 +315,7 @@ void ECLPackerModule::event()
             // pack time
             tim = m_eclDigits[i_digit]->getTimeFit();
           }
-          unsigned int hit_data = ((qua & 3) << 30) & 0xC0000000;
+          unsigned int hit_data = ((unsigned int)(qua & 3) << 30) & 0xC0000000;
           hit_data |= (tim & 0xFFF) << 18;
           hit_data |= ((amp + 128) & 0x3FFFF);
           buff[iFINESSE].push_back(hit_data);
